+++ packages/nc-gui/components/project/spreadsheet/components/editColumn/RatingOptions.vue
