+++ packages/nc-gui/components/project/spreadsheet/components/editableCell/RatingCell.vue
